Move branch storage to utils package
Create utils package and move everything related to branch creation into
the utils.go file. Then call `utils.Store` to store a branch.

Doing this so I can easily store a branch in history from a
`git-checkout-history` call. This is because we want to save to history
even if you're checking out using the history shorthand.